add DetectSign Class
