+++ .gitignore
@@ -2,3 +2,4 @@ venv
 .idea/
 *.pyc
 *.pyo
+server/logs
