--- ASL_Classifier/Pipfile
@@ -1,15 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-tensorflow = "==1.2.1"
-matplotlib = "==2.0.2"
-numpy = "==1.13.0"
-opencv-python = "==3.2.0.8"
-
-[requires]
-python_version = "3.5"
