+++ __init__.py
@@ -1,0 +1,3 @@
+import src
+
+__all__ = ["src"]
